+++ src/validations/prefab/linkedOption.ts
@@ -15,6 +15,7 @@ export const linkedOptionConfigurationSchema = Joi.object({
     Joi.object({
       name: Joi.string(),
       value: Joi.alternatives().try(Joi.boolean(), Joi.string(), Joi.number()),
+      icon: Joi.string(),
     }),
   ),
   condition: Joi.object({
@@ -31,4 +32,6 @@ export const linkedOptionSchema = Joi.object({
   type: Joi.string().valid('LINKED_OPTION').required(),
   value: linkedOptionValueSchema,
   configuration: linkedOptionConfigurationSchema,
+  showInAddChild: Joi.boolean(),
+  showInReconfigure: Joi.boolean(),
 });